Extract export stuff to a different component
